python-rag-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
llama-index-core==0.10.11
llama-index-vector-stores-supabase==0.1.3
llama-index-embeddings-openai==0.1.6
llama-index-readers-file==0.1.4
supabase==2.0.2
python-multipart==0.0.6
openai==1.3.5
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pydantic==2.5.0 